Update ptud46_controller_node.cpp
do i really need that sendCmd function for the laser scanner? trying to find a efficient way to tackle this.
and do we actually define the max_ang and min_ang as 30 and -30 or does it automatically take that given that's the max our laser scanner can go?
I have to change the variable names too.